requirements.txt
pandas==2.2.2
numpy==1.26.4
yfinance==0.2.40
python-dotenv==1.0.1
streamlit==1.38.0
requests==2.32.3
polars==1.6.0
pydantic==2.8.2
joblib==1.4.2
